Extract evm decorations into separate classes
