+++ UnstoppableWallet/UnstoppableWallet/Modules/SendNew/EvmSendHandler.swift
@@ -11,59 +11,13 @@ class EvmSendHandler {
     let baseToken: Token
     let transactionData: TransactionData
     let evmKitWrapper: EvmKitWrapper
+    let decorator = EvmDecorator()
     let evmFeeEstimator = EvmFeeEstimator()
 
     init(baseToken: Token, transactionData: TransactionData, evmKitWrapper: EvmKitWrapper) {
         self.baseToken = baseToken
         self.transactionData = transactionData
         self.evmKitWrapper = evmKitWrapper
-    }
-
-    private func decorate(transactionData: TransactionData, transactionDecoration: TransactionDecoration?) -> Decoration {
-        var type: Decoration.`Type`?
-        var customSendButtonTitle: String?
-
-        switch transactionDecoration {
-        case let decoration as OutgoingDecoration:
-            type = .outgoingEvm(
-                to: decoration.to,
-                value: baseToken.decimalValue(value: decoration.value)
-            )
-
-        case let decoration as OutgoingEip20Decoration:
-            if let token = try? coinManager.token(query: .init(blockchainType: baseToken.blockchainType, tokenType: .eip20(address: decoration.contractAddress.hex))) {
-                type = .outgoingEip20(
-                    to: decoration.to,
-                    value: token.decimalValue(value: decoration.value),
-                    token: token
-                )
-            }
-
-        case let decoration as ApproveEip20Decoration:
-            if let token = try? coinManager.token(query: .init(blockchainType: baseToken.blockchainType, tokenType: .eip20(address: decoration.contractAddress.hex))) {
-                type = .approveEip20(
-                    spender: decoration.spender,
-                    value: token.decimalValue(value: decoration.value),
-                    token: token
-                )
-
-                let isRevoke = decoration.value == 0
-
-                customSendButtonTitle = isRevoke ? "send.confirmation.slide_to_revoke".localized : "send.confirmation.slide_to_approve".localized
-            }
-        default:
-            ()
-        }
-
-        return Decoration(
-            type: type ?? .unknown(
-                to: transactionData.to,
-                value: baseToken.decimalValue(value: transactionData.value),
-                input: transactionData.input,
-                method: evmLabelManager.methodLabel(input: transactionData.input)
-            ),
-            customSendButtonTitle: customSendButtonTitle
-        )
     }
 }
 
@@ -103,7 +57,7 @@ extension EvmSendHandler: ISendHandler {
         }
 
         let transactionDecoration = evmKitWrapper.evmKit.decorate(transactionData: transactionData)
-        let decoration = decorate(transactionData: transactionData, transactionDecoration: transactionDecoration)
+        let decoration = decorator.decorate(baseToken: baseToken, transactionData: transactionData, transactionDecoration: transactionDecoration)
 
         return SendData(
             decoration: decoration,
@@ -143,11 +97,11 @@ extension EvmSendHandler: ISendHandler {
 
 extension EvmSendHandler {
     class SendData: BaseSendEvmData, ISendData {
-        let decoration: Decoration
+        let decoration: EvmDecoration
         let transactionData: TransactionData?
         let transactionError: Error?
 
-        init(decoration: Decoration, transactionData: TransactionData?, transactionError: Error?, gasPrice: GasPrice?, evmFeeData: EvmFeeData?, nonce: Int?) {
+        init(decoration: EvmDecoration, transactionData: TransactionData?, transactionError: Error?, gasPrice: GasPrice?, evmFeeData: EvmFeeData?, nonce: Int?) {
             self.decoration = decoration
             self.transactionData = transactionData
             self.transactionError = transactionError
@@ -182,18 +136,7 @@ extension EvmSendHandler {
         }
 
         func sections(baseToken: Token, currency: Currency, rates: [String: Decimal]) -> [[SendField]] {
-            var sections: [[SendField]]
-
-            switch decoration.type {
-            case let .outgoingEvm(to, value):
-                sections = outgoingSections(token: baseToken, to: to, value: value, currency: currency, rates: rates)
-            case let .outgoingEip20(to, value, token):
-                sections = outgoingSections(token: token, to: to, value: value, currency: currency, rates: rates)
-            case let .approveEip20(spender, value, token):
-                sections = approveSections(token: token, spender: spender, value: value, currency: currency, rates: rates)
-            case let .unknown(to, value, input, method):
-                sections = unknownSections(baseToken: baseToken, to: to, value: value, input: input, method: method, currency: currency, rates: rates)
-            }
+            var sections = decoration.sections(baseToken: baseToken, currency: currency, rates: rates)
 
             if let nonce {
                 sections.append(
@@ -207,119 +150,10 @@ extension EvmSendHandler {
 
             return sections
         }
-
-        private func outgoingSections(token: Token, to: EvmKit.Address, value: Decimal, currency: Currency, rates: [String: Decimal]) -> [[SendField]] {
-            [
-                [
-                    amountField(
-                        title: "send.confirmation.you_send".localized,
-                        token: token,
-                        value: value,
-                        currency: currency,
-                        rate: rates[token.coin.uid],
-                        type: .neutral
-                    ),
-                    .address(
-                        title: "send.confirmation.to".localized,
-                        value: to.eip55,
-                        blockchainType: token.blockchainType
-                    ),
-                ],
-            ]
-        }
-
-        private func approveSections(token: Token, spender: EvmKit.Address, value: Decimal, currency: Currency, rates: [String: Decimal]) -> [[SendField]] {
-            let isRevokeAllowance = value == 0 // Check approved new value or revoked last allowance
-
-            let amountField: SendField
-
-            if isRevokeAllowance {
-                amountField = .amount(
-                    title: "approve.confirmation.you_revoke".localized,
-                    token: token,
-                    coinValueType: .withoutAmount(kind: .token(token: token)),
-                    currencyValue: nil,
-                    type: .neutral
-                )
-            } else {
-                amountField = self.amountField(
-                    title: "approve.confirmation.you_approve".localized,
-                    token: token,
-                    value: value,
-                    currency: currency,
-                    rate: rates[token.coin.uid],
-                    type: .neutral
-                )
-            }
-
-            return [
-                [
-                    amountField,
-                    .address(
-                        title: "approve.confirmation.spender".localized,
-                        value: spender.eip55,
-                        blockchainType: token.blockchainType
-                    ),
-                ],
-            ]
-        }
-
-        private func unknownSections(baseToken: Token, to: EvmKit.Address, value: Decimal, input _: Data, method _: String?, currency: Currency, rates: [String: Decimal]) -> [[SendField]] {
-            [
-                [
-                    amountField(
-                        title: "send.confirmation.transfer".localized,
-                        token: baseToken,
-                        value: value,
-                        currency: currency,
-                        rate: rates[baseToken.coin.uid],
-                        type: .neutral
-                    ),
-                    .address(
-                        title: "send.confirmation.to".localized,
-                        value: to.eip55,
-                        blockchainType: baseToken.blockchainType
-                    ),
-                    // TODO: show input and method
-                ],
-            ]
-        }
-
-        private func amountField(title: String, token: Token, value: Decimal, currency: Currency, rate: Decimal?, type: SendField.AmountType) -> SendField {
-            let coinValue = CoinValue(kind: .token(token: token), value: Decimal(sign: type.sign, exponent: value.exponent, significand: value.significand))
-
-            return .amount(
-                title: title,
-                token: token,
-                coinValueType: coinValue.isMaxValue ? .infinity(kind: coinValue.kind) : .regular(coinValue: coinValue),
-                currencyValue: coinValue.isMaxValue ? nil : rate.map { CurrencyValue(currency: currency, value: $0 * value) },
-                type: type
-            )
-        }
     }
 }
 
 extension EvmSendHandler {
-    struct Decoration {
-        let type: Type
-        let customSendButtonTitle: String?
-
-        enum `Type` {
-            case outgoingEvm(to: EvmKit.Address, value: Decimal)
-            case outgoingEip20(to: EvmKit.Address, value: Decimal, token: Token)
-            case approveEip20(spender: EvmKit.Address, value: Decimal, token: Token)
-            case unknown(to: EvmKit.Address, value: Decimal, input: Data, method: String?)
-        }
-
-        var rateCoins: [Coin] {
-            switch type {
-            case let .outgoingEip20(_, _, token): return [token.coin]
-            case let .approveEip20(_, _, token): return [token.coin]
-            default: return []
-            }
-        }
-    }
-
     enum SendError: Error {
         case invalidData
         case noGasPrice
